refactor(register): 이미지 인풋 리팩토링 및 form 연결 해제
